[ANI-21-spacing-block] refactor: 다른 컴포넌트의 영향을 받지 않게 하기 위해 파일 위치를 수정합니다.
