+++ packages/ui/src/components/atom/radio-group/RadioGroup.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import RadioGroup from './RadioGroup'
+import RadioGroup from '.'
 
 const meta = {
   title: 'Atom/RadioGroup',
