+++ packages/ui/src/components/atom/typography/index.tsx
@@ -34,7 +34,7 @@ export const typographyVariants = cva('', {
       [TypographyColors.NEUTRAL]: 'text-label-neutral',
       [TypographyColors.ALTERNATIVE]: 'text-label-alternative',
       [TypographyColors.ASSISTIVE]: 'text-label-assistive',
-      [TypographyColors.DISABLED]: 'text-label-disabled'
+      [TypographyColors.DISABLE]: 'text-label-disable'
     }
   },
   defaultVariants: {
