+++ packages/ui/src/components/renderComponents.tsx
@@ -1,6 +1,7 @@
 import { type ServerDrivenComponentType } from '@/types/server-driven.js'
 
-import RadioGroup from './atom/radio-group/RadioGroup.js'
+import RadioGroup from './atom/radio-group'
+import Typography from './atom/typography'
 
 export function renderComponents(componentData: ServerDrivenComponentType) {
   const { type, props, children } = componentData
@@ -17,6 +18,9 @@ export function renderComponents(componentData: ServerDrivenComponentType) {
           {props.children}
         </RadioGroup.Item>
       )
+
+    case 'Typography':
+      return <Typography {...props}>{props.children}</Typography>
 
     default:
       console.warn(`Unknown component type: ${type}`)
