--- packages/ui/src/index.ts
@@ -1,2 +1,0 @@
-export { default as RadioGroup } from '@/components/atom/radio-group/RadioGroup'
-export { default as Typography } from '@/components/atom/typography/Typography'
